backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
google-genai>=0.3.0
langfuse>=2.60.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.24.0
